backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
anthropic==0.43.0
redis[hiredis]==5.0.1
python-hcl2==4.3.2
gitpython==3.1.41
python-multipart==0.0.9
weasyprint==60.2
jinja2==3.1.3
cryptography==42.0.2
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
